Техническое задание проекта №12

+++ src/main/java/ru/yandex/practicum/filmorate/storage/catalog/CatalogDbStorage.java
@@ -47,14 +47,14 @@ public class CatalogDbStorage implements CatalogStorage {
 
     private Genre mapGenre(ResultSet rs, int rowNum) throws SQLException {
         return Genre.builder()
-                .id(rs.getInt("id"))
+                .id(rs.getLong("id"))
                 .name(rs.getString("name"))
                 .build();
     }
 
     private Mpa mapMpa(ResultSet rs, int rowNum) throws SQLException {
         return Mpa.builder()
-                .id(rs.getInt("id"))
+                .id(rs.getLong("id"))
                 .name(rs.getString("name"))
                 .build();
     }
